Write R object with corresponding stamp (hash digest)

The st_write function is intended to be a wrapper of several other functions in different packages that save data disk. Yet, it goes several steps beyond that. First it creates the stamps (e.g., hashes) as part of the attributes of the object and saves it in a different file for easy access. Also, it may create vintage files of the object to keep track of changes. The philosophy of this package is increase speed in processes that work with many files, avoiding the need to load files to check whether the data has changed or not. Since disk space is cheap and time is not, stamp may be redundant in the files it saves.

Usage

st_write(
  x,
  file,
  ext = fs::path_ext(file),
  st_dir = NULL,
  st_name = NULL,
  complete_stamp = getOption("stamp.completestamp"),
  recurse = FALSE,
  force = FALSE,
  algo = getOption("stamp.digest.algo"),
  vintage = getOption("stamp.vintage"),
  vintage_dir = NULL,
  verbose = getOption("stamp.verbose"),
  ...
)

Arguments

x

R object to write to disk as per limitations of file format.

file

character: File or connection to write to

ext

character: format or extension of file. Default is fs::path_ext(file)

st_dir

character: Directory to store stamp files. By default it is a subdirectory at the same level of file.

st_name

character: name of stamp in file. All stamp files are prefixed with value in option "stamp.stamp_prefix", which by default is "st_". You don't need to add the prefix.

complete_stamp

logical: Whether to add a complete report of data.frame to stamp file. You need the skimr package. If skimr is not in namespace, limited but lighter report will be added.

recurse

logical: is TRUE if directory in file does not it will be created. Default is FALSE

force

logical: replace file in disk even if has hasn't changed

algo

character: Algorithm to be used in digest::digest(). Default is "sha1"

vintage

logical: Whether to save vintage versions x. Default TRUE

vintage_dir

character: Directory to save vintages of x. By default it is a subdirectory at the same level of file

verbose

logical: whether to display additional information. This could be changed in option "stamp.verbose". Default is TRUE

...

not is used right now

Value

TRUE is object was saved successfully. FALSE otherwise.

Details

Object x is stored in file but its hash (i.e., stamp) is stored in subdirectory st_file.

Vintage files Vintage files are optional but play an important role for

replicability purposes. We highly recommend you turn this option off if you don't have enough space in your disk.

Examples

if (FALSE) {
  tfile <- fs::file_temp(ext = "qs")
  st_write(df, tfile)
}