Confirm stamp has not changed

verifies that, were the stamp recalculated, it would match the one previously set with stamp_set() of stamp_save().

Usage

stamp_confirm(
  x,
  using = c("self", "stamp"),
  st_dir = NULL,
  st_name = NULL,
  st_ext = getOption("stamp.default.ext"),
  verbose = getOption("stamp.verbose"),
  ...
)

Arguments

x

An arbitrary R object which will then be passed to the base::serialize function

using

character: either "self" of "stamp" (see details).

st_dir

character: parent directory where the stamp file if saved.

st_name

character: name of stamp in file. All stamp files are prefixed with the value in option "stamp.stamp_prefix", which by default is "st". You don't need to add the prefix

st_ext

character: format of stamp file to save. Default is value in option "stamp.default.ext"

verbose

logical: whether to display additional information. This could be changed in option "stamp.verbose". Default is TRUE

...

Arguments passed on to stamp_get

algo

character: default is value in option "stamp.digest.algo". This argument is the algorithms to be used; currently available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64, murmur32, spookyhash and blake3

serialize

A logical variable indicating whether the object should be serialized using serialize (in ASCII form). Setting this to FALSE allows to compare the digest output of given character strings to known control output. It also allows the use of raw vectors such as the output of non-ASCII serialization.

file

A logical variable indicating whether the object is a file name or a file name if object is not specified.

length

Number of characters to process. By default, when length is set to Inf, the whole string or file is processed.

skip

Number of input bytes to skip before calculating the digest. Negative values are invalid and currently treated as zero. Special value "auto" will cause serialization header to be skipped if serialize is set to TRUE (the serialization header contains the R version number thus skipping it allows the comparison of hashes across platforms and some R versions).

ascii

This flag is passed to the serialize function if serialize is set to TRUE, determining whether the hash is computed on the ASCII or binary representation.

raw

A logical variable with a default value of FALSE, implying digest returns digest output as ASCII hex values. Set to TRUE to return digest output in raw (binary) form. Note that this option is supported by most but not all of the implemented hashing algorithms

seed

an integer to seed the random number generator. This is only used in the xxhash32, xxhash64 and murmur32 functions and can be used to generate additional hashes for the same input if desired.

errormode

A character value denoting a choice for the behaviour in the case of error: ‘stop’ aborts (and is the default value), ‘warn’ emits a warning and returns NULL and ‘silent’ suppresses the error and returns an empty string.

Value

Logical value. FALSE if the objects do not match and TRUE if they do.

Details

using If "self" verifies that stamp has not changed by recalculating it and comparing with the one previously set x. If no stamp has been set in x, it yields error. If "stamp", you must provide a st_dir and a st_name, which correspond to the stamp in file to compare with. If only st_dir is provided, it is assumes to contain the name of the file as well. Otherwise, error is yield.

See also