Add attributes and characteristics of x to be used in stamp file

Add attributes and characteristics of x to be used in stamp file

Usage

stamp_x_attr(x, complete_stamp = getOption("stamp.completestamp"))

Arguments

x

R object to write to disk as per limitations of file format.

complete_stamp

logical: Whether to add a complete report of data.frame to stamp file. You need the skimr package. If skimr is not in namespace, limited but lighter report will be added.

Value

list of attributes

See also

Examples

x <- data.frame(a = 1:10, b = letters[1:10])
stamp_x_attr(x)
#> $names
#> [1] "a" "b"
#> 
#> $class
#> [1] "data.frame"
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $skim
#> ── Data Summary ────────────────────────
#>                            Values
#> Name                       x     
#> Number of rows             10    
#> Number of columns          2     
#> _______________________          
#> Column type frequency:           
#>   character                1     
#>   numeric                  1     
#> ________________________         
#> Group variables            None  
#> 
#> ── Variable type: character ────────────────────────────────────────────────────
#>   skim_variable n_missing complete_rate min max empty n_unique whitespace
#> 1 b                     0             1   1   1     0       10          0
#> 
#> ── Variable type: numeric ──────────────────────────────────────────────────────
#>   skim_variable n_missing complete_rate mean   sd p0  p25 p50  p75 p100 hist 
#> 1 a                     0             1  5.5 3.03  1 3.25 5.5 7.75   10 ▇▇▇▇▇
#> 
#> $type
#> [1] "list"
#>