Set an attribute stamp to R object

This functions does the same as stamp_get() but stores the stamps as an attribute in the object. If the object is not saved afterward the stamps won't be permanent. Yet, it is useful for quick verification.

Usage

stamp_set(
  x,
  st_name = NULL,
  verbose = getOption("stamp.verbose"),
  replace = FALSE,
  ...
)

stamp_call(st_name)

Arguments

x

An arbitrary R object which will then be passed to the base::serialize function

st_name

character: Name of stamp to be saved in .stamp env.

verbose

logical: whether to display additional information. This could be changed in option "stamp.verbose". Default is TRUE

replace

Logical: if TRUE and st_name already exists in .stamp environment, it will be replaced with new stamp. If FALSE it gives an error. Default is FALSE

...

Arguments passed on to stamp_get

algo

character: default is value in option "stamp.digest.algo". This argument is the algorithms to be used; currently available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64, murmur32, spookyhash and blake3

serialize

A logical variable indicating whether the object should be serialized using serialize (in ASCII form). Setting this to FALSE allows to compare the digest output of given character strings to known control output. It also allows the use of raw vectors such as the output of non-ASCII serialization.

file

A logical variable indicating whether the object is a file name or a file name if object is not specified.

length

Number of characters to process. By default, when length is set to Inf, the whole string or file is processed.

skip

Number of input bytes to skip before calculating the digest. Negative values are invalid and currently treated as zero. Special value "auto" will cause serialization header to be skipped if serialize is set to TRUE (the serialization header contains the R version number thus skipping it allows the comparison of hashes across platforms and some R versions).

ascii

This flag is passed to the serialize function if serialize is set to TRUE, determining whether the hash is computed on the ASCII or binary representation.

raw

A logical variable with a default value of FALSE, implying digest returns digest output as ASCII hex values. Set to TRUE to return digest output in raw (binary) form. Note that this option is supported by most but not all of the implemented hashing algorithms

seed

an integer to seed the random number generator. This is only used in the xxhash32, xxhash64 and murmur32 functions and can be used to generate additional hashes for the same input if desired.

errormode

A character value denoting a choice for the behaviour in the case of error: ‘stop’ aborts (and is the default value), ‘warn’ emits a warning and returns NULL and ‘silent’ suppresses the error and returns an empty string.

Value

invisible stamp from stamp_get() but it can now be called with stamp_call()

list with stamp values

See also

Examples

x <- data.frame(a = 1:10, b = letters[1:10])
stamp_set(x, st_name = "sttest")
stamp_call("sttest")
#> $stamps
#> $stamps$a
#> [1] "1f13173d1d78e38b1df3ddeff0ae3b8efd9c20b8"
#> 
#> $stamps$b
#> [1] "52564269a73de74842708a3aec99783949834294"
#> 
#> 
#> $time
#> $time$tz
#> [1] "UTC"
#> 
#> $time$tformat
#> [1] "%Y%m%d%H%M%S"
#> 
#> $time$usetz
#> [1] FALSE
#> 
#> $time$st_time
#> [1] "20230121051626"
#> 
#> 
#> $algo
#> [1] "sha1"
#>